backend/requirements.txt
fastapi
asyncpg
asyncio
uvicorn
requests
SQLAlchemy
psycopg2-binary
elasticsearch
python-dotenv
sentence-transformers
torch
openai-whisper
opencv-python
pillow
ollama
pydantic
tenacity
starlette
moviepy==1.0.3
python-multipart